Reorganize onboarding screen structure
