+++ lib/app.dart
@@ -9,15 +9,15 @@ import 'src/core/motion/onboarding_motion.dart';
 import 'src/core/theme/app_theme.dart';
 import 'src/core/theme/legacy_material_theme.dart';
 import 'src/features/home/screens/home_screen.dart';
-import 'src/features/onboarding/screens/address_types_screen.dart';
-import 'src/features/onboarding/screens/create_wallet_screen.dart';
+import 'src/features/onboarding/create/address_types_screen.dart';
+import 'src/features/onboarding/create/create_wallet_screen.dart';
+import 'src/features/onboarding/create/intro_zcash_screen.dart';
+import 'src/features/onboarding/create/onboarding_split_view.dart';
+import 'src/features/onboarding/create/secret_passphrase_screen.dart';
+import 'src/features/onboarding/create/things_to_know_screen.dart';
 import 'src/features/onboarding/screens/import_wallet_screen.dart';
-import 'src/features/onboarding/screens/intro_zcash_screen.dart';
-import 'src/features/onboarding/screens/onboarding_split_view.dart';
-import 'src/features/onboarding/screens/set_password_screen.dart';
-import 'src/features/onboarding/screens/secret_passphrase_screen.dart';
-import 'src/features/onboarding/screens/things_to_know_screen.dart';
-import 'src/features/onboarding/screens/unlock_screen.dart';
+import 'src/features/onboarding/shared/set_password_screen.dart';
+import 'src/features/onboarding/unlock_screen.dart';
 import 'src/features/onboarding/welcome.dart';
 import 'src/features/history/screens/history_screen.dart';
 import 'src/features/receive/screens/receive_screen.dart';
